requirements.txt
openai
langchain
langchain_community
pypdf2
slack_sdk
python-dotenv
chromadb
tiktoken